requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
aiohttp=3.6.2=pypi_0
appdirs=1.4.3=py36h28b3542_0
async-timeout=3.0.1=pypi_0
attrs=19.3.0=py_0
black=19.10b0=py_0
blas=1.0=mkl
ca-certificates=2020.1.1=0
certifi=2019.11.28=py36_0
chardet=3.0.4=pypi_0
click=7.0=py36_0
cpuonly=1.0=0
fastcache=1.1.0=py36h7b6447c_0
gmp=6.1.2=h6c8ec71_1
gmpy2=2.0.8=py36h10f8cd9_2
idna=2.9=pypi_0
idna-ssl=1.1.0=pypi_0
intel-openmp=2020.0=166
ld_impl_linux-64=2.33.1=h53a641e_7
libedit=3.1.20181209=hc058e9b_0
libffi=3.2.1=hd88cf55_4
libgcc-ng=9.1.0=hdf63c60_0
libgfortran-ng=7.3.0=hdf63c60_0
libstdcxx-ng=9.1.0=hdf63c60_0
mkl=2020.0=166
mkl-service=2.3.0=py36he904b0f_0
mkl_fft=1.0.15=py36ha843d7b_0
mkl_random=1.1.0=py36hd6b4f25_0
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.1=hdf1c602_3
mpmath=1.1.0=py36_0
multidict=4.7.5=pypi_0
mypy_extensions=0.4.3=py36_0
ncurses=6.2=he6710b0_0
numpy=1.18.1=py36h4f9e942_0
numpy-base=1.18.1=py36hde5b4d6_1
openssl=1.1.1d=h7b6447c_4
pathspec=0.7.0=py_0
pip=20.0.2=py36_1
python=3.6.10=h0371630_0
readline=7.0=h7b6447c_5
regex=2020.2.20=py36h7b6447c_0
setuptools=45.2.0=py36_0
six=1.14.0=py36_0
slackclient=2.5.0=pypi_0
sqlite=3.31.1=h7b6447c_0
sympy=1.5.1=py36_0
tk=8.6.8=hbc83047_0
toml=0.10.0=py36h28b3542_0
typed-ast=1.4.1=py36h7b6447c_0
typing_extensions=3.7.4.1=py36_0
wheel=0.34.2=py36_0
xz=5.2.4=h14c3975_4
yarl=1.4.2=pypi_0
zlib=1.2.11=h7b6447c_3
